[feature]: Add clsx to handle user guess state styles
Introduce clsx library to manage the dynamic styles of user guesses. User guesses are now visually differentiated based on their accuracy in both letter position and existence within the word. Green color is applied if the guessed letter is in the correct position and matches the letter in the daily word. Yellow color is applied if the guessed letter exists in the daily word but is not in the correct position. Additionally, a constant.js file has been created to simulate the dailyWord for testing purposes.